pyproject.toml
[project]
name = "mcp-youtrack"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for YouTrack integration."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.13"

dependencies = [
     "mcp[cli]>=1.3.0",
     "python-dotenv>=1.0.1",
     "uvicorn>=0.34.0",
     "youtrack-sdk @ git+https://github.com/moneymeets/youtrack-sdk.git"
]

[project.scripts]
mcp-youtrack = "mcp_youtrack.main:main"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest"
]

[tool.hatch.build.targets.wheel]
packages = ["mcp_youtrack"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
